requirements.txt
httpx>=0.27.0
pydantic>=2.6.0
PyYAML>=6.0.1
websockets>=12.0
streamlit>=1.32.0
pandas>=2.2.0
pytest>=8.0.0
rich>=13.7.0
